MIN-11-Entity-Seat
ManyToOne theater_id to seat
