--- src/main/java/org/example/kinoxpbackend/Category.java
@@ -1,27 +1,0 @@
-package org.example.kinoxpbackend;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-
-import java.time.LocalDateTime;
-@Getter
-@Setter
-@NoArgsConstructor
-@Entity
-public class Category {
-    @jakarta.persistence.Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-
-    private int Id;
-   @Column(nullable = false, unique = true)
-    private String name;
-    @CreationTimestamp
-    private LocalDateTime created;
-    @UpdateTimestamp
-    private LocalDateTime updated;
-
-}
